requirements.txt
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.7
django-allauth==0.45.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==8.12.1
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
model-bakery==1.3.2
oauthlib==3.1.1
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
